3_ML/requirements.txt
azure-functions
pandas
numpy
ujson
azure.cognitiveservices.language.textanalytics
msrest
azure-search-documents==1.0.0b3
